Rearrange the tests, so they run sanity test and check with mypy
